+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-05-26 21:36:06
-// Commit: fix(security): prevent double tap duplicate action in security
+// Updated at: 2026-05-27 11:11:10
+// Commit: feat(security): add async security stream provider
 
 class securityComponentState {
-  final String commitRef = '1477';
-  final String lastMessage = 'fix(security): prevent double tap duplicate action in security';
-  final DateTime timestamp = DateTime.parse('2026-05-26T21:36:06+05:00');
+  final String commitRef = '1486';
+  final String lastMessage = 'feat(security): add async security stream provider';
+  final DateTime timestamp = DateTime.parse('2026-05-27T11:11:10+05:00');
 }